requirements.txt
asgiref==3.4.1
bleach==4.1.0
dj-database-url==0.5.0
Django==3.2.8
django-ckeditor==6.1.0
django-easy-validator==1.7.0
django-heroku==0.0.0
django-js-asset==1.2.2
django-richtextfield==1.6.1
django-summernote==0.8.20.0
fontawesome-free==5.15.4
gunicorn==20.1.0
mysqlclient==2.0.3
packaging==21.2
Pillow==8.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyparsing==2.4.7
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
whitenoise==5.3.0
